requirements.txt
# --- Core ---
torch>=2.0.0
torchvision
numpy
pillow

# --- FastAPI App ---
fastapi
uvicorn

# --- TorchServe ---
torchserve
torch-model-archiver

# --- MLflow 관련 (모델 추적 등) ---
mlflow

# --- 기타 유틸 (선택) ---
python-multipart
nvgpu